\name{DJ.col}
\alias{DJ.col}
\title{DJ col}
\description{
}
\usage{
DJ.col(cl, w, gray.l = FALSE)
}
\arguments{
  \item{cl}{
}
  \item{w}{
}
  \item{gray.l}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
