\name{gen.op.set}
\alias{gen.op.set}
\title{gen op set}
\description{
}
\usage{
gen.op.set(param, assign.op = TRUE, p.env = parent.frame())
}
\arguments{
  \item{param}{
}
  \item{assign.op}{
}
  \item{p.env}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
