\name{inv.par}
\alias{inv.par}
\title{inv par}

\usage{
inv.par(ic = TRUE, par.opt = c("bg", "col", "col.axis", "col.lab", "col.main", "col.sub", "fg"), ...)
}
\arguments{
  \item{ic}{
}
  \item{par.opt}{
}
  \item{\dots}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





