\name{par.op.set}
\alias{par.op.set}
\title{par op set}
\description{
}
\usage{
par.op.set(param, assign.op = TRUE, p.env = parent.frame())
}
\arguments{
  \item{param}{
}
  \item{assign.op}{
}
  \item{p.env}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
