% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture-package.R
\docType{package}
\name{soiltexture-package}
\alias{soiltexture-package}
\alias{soiltexture}
\title{c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_title(\"#1\")}",
"soiltexture")\Sexpr{tools:::Rd_package_title("soiltexture")}}
\description{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_description(\"#1\")}",
"soiltexture")\Sexpr{tools:::Rd_package_description("soiltexture")}
}
\details{
The DESCRIPTION file:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_DESCRIPTION(\"#1\")}",
"soiltexture")\Sexpr{tools:::Rd_package_DESCRIPTION("soiltexture")}
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_indices(\"#1\")}",
"soiltexture")\Sexpr{tools:::Rd_package_indices("soiltexture")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/julienmoeys/soiltexture}
  \item Report bugs at \url{https://github.com/julienmoeys/soiltexture/issues}
}

}
\author{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_author(\"#1\")}",
"soiltexture")\Sexpr{tools:::Rd_package_author("soiltexture")}

Maintainer:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_maintainer(\"#1\")}",
"soiltexture")\Sexpr{tools:::Rd_package_maintainer("soiltexture")}
}
\keyword{package}
