\name{TT.auto.set}
\alias{TT.auto.set}
\title{Internal. Retrieve and set default values for parameters (par() or not), when NULL.}
\description{Retrieve and set default values for parameters (par() or not), when NULL.}
\usage{TT.auto.set(fun = sys.function(which = -1), assign.op = TRUE, 
    p.env = parent.frame(), set.par = TRUE)}
\arguments{
  \item{fun}{
}
  \item{assign.op}{
}
  \item{p.env}{
}
  \item{set.par}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





