\name{TT.axis.arrows}
\alias{TT.axis.arrows}
\title{Internal. Plot the axis' arrows of a texture triangle plot. }
\description{Plot the axis' arrows of a texture triangle plot. }
\usage{
TT.axis.arrows(geo,
css.lab = NULL,
a.l = TT.get("arrows.lims"),
a.h.s = TT.get("arrows.head.shift"),
a.t.s = TT.get("arrows.text.shift"),
a.t.s2 = TT.get("arrows.text.shift2"),
a.b.s = TT.get("arrows.base.shift"),
text.tol = NULL,
text.sum = NULL,
blr.clock = NULL,
tlr.an = NULL,
base.css.ps.lim = NULL,
tri.sum.tst = FALSE,
tri.pos.tst = FALSE,
lwd.lab = NULL,
arrows.lty = NULL,
col.lab = NULL,
font.lab = NULL,
cex.lab = NULL,
family.op = NULL,
unit.ps = NULL,
unit.tx = NULL,
lang = NULL)
}
\arguments{
  \item{geo}{
}
  \item{css.lab}{
}
  \item{a.l}{
}
  \item{a.h.s}{
}
  \item{a.t.s}{
}
  \item{a.t.s2}{
}
  \item{a.b.s}{
}
  \item{text.tol}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
  \item{tlr.an}{
}
  \item{base.css.ps.lim}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{lwd.lab}{
}
  \item{arrows.lty}{
}
  \item{col.lab}{
}
  \item{font.lab}{
}
  \item{cex.lab}{
}
  \item{family.op}{
}
  \item{unit.ps}{
}
  \item{unit.tx}{
}
  \item{lang}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.}





