\name{TT.baseplot}
\alias{TT.baseplot}
\title{Internal. Create an empty plot scene for a texture triangle.}
\description{Create an empty plot where a texture triangle can be drawn with 
other secondary functions (frame, axis, ...). Also return the 
'geo' parameters needed by these secondary functions.}
\usage{TT.baseplot(geo = NULL, class.sys = "none", blr.clock = NULL, 
    tlr.an = NULL, blr.tx = NULL, text.sum = NULL, 
    base.css.ps.lim = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, 
    text.tol = NULL, unit.ps = NULL, unit.tx = NULL, 
    b.lim = NULL, l.lim = NULL, main = NULL, new.mar = NULL, 
    bg = NULL, fg = NULL, col = NULL, cex.main = NULL, 
    lang = NULL)}
\arguments{
  \item{geo}{
}
  \item{class.sys}{
}
  \item{blr.clock}{
}
  \item{tlr.an}{
}
  \item{blr.tx}{
}
  \item{text.sum}{
}
  \item{base.css.ps.lim}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{text.tol}{
}
  \item{unit.ps}{
}
  \item{unit.tx}{
}
  \item{b.lim}{
}
  \item{l.lim}{
}
  \item{main}{
}
  \item{new.mar}{
}
  \item{bg}{
}
  \item{fg}{
}
  \item{col}{
}
  \item{cex.main}{
}
  \item{lang}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





