\name{TT.blr.ps.lim}
\alias{TT.blr.ps.lim}
\title{TT blr ps lim}
\description{
}
\usage{
TT.blr.ps.lim(blr.tx, css.ps.lim)
}
\arguments{
  \item{blr.tx}{
}
  \item{css.ps.lim}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
