\name{TT.blr.tx.check}
\alias{TT.blr.tx.check}
\title{TT blr tx check}
\description{
}
\usage{
TT.blr.tx.check(blr.tx, css.names)
}
\arguments{
  \item{blr.tx}{
}
  \item{css.names}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
