\name{TT.check.ps.lim}
\alias{TT.check.ps.lim}
\title{Internal. Check the consistency between 'base.ps.lim' and 'dat.ps.lim'. }
\description{Check the consistency between 'base.ps.lim' and 'dat.ps.lim'. 
5 tests performed.}
\usage{TT.check.ps.lim(base.ps.lim, dat.ps.lim, ps.lim.length = c(4, 
    4))}
\arguments{
  \item{base.ps.lim}{
}
  \item{dat.ps.lim}{
}
  \item{ps.lim.length}{vector of 2 integers. Number of particle size classes + 1. c(base,dat)}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





