\name{TT.chemometrics.alr}
\alias{TT.chemometrics.alr}
\title{TT chemometrics alr}

\usage{
TT.chemometrics.alr(X, divisorvar, css.names)
}
\arguments{
  \item{X}{
}
  \item{divisorvar}{
}
  \item{css.names}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





