\name{TT.classes.tbl}
\alias{TT.classes.tbl}
\title{TT classes tbl}
\description{
}
\usage{
TT.classes.tbl(class.sys = "FAO50.TT", collapse = ", ")
}
\arguments{
  \item{class.sys}{
}
  \item{collapse}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
