\name{TT.col2hsv}
\alias{TT.col2hsv}
\title{Convert any colors to hsv. }
\description{Convert any colors to hsv. Wrapper around rgb2hsv() and col2rgb(). }
\usage{
TT.col2hsv(col)
}
\arguments{
  \item{col}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.}





