% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.contour}
\alias{TT.contour}
\title{Wrapper for the contour() function adapted to texture triangles.}
\usage{
TT.contour(
  geo,
  x,
  add = FALSE,
  tri.sum.tst = NULL,
  tri.pos.tst = NULL,
  text.tol = NULL,
  unit.ps = NULL,
  unit.tx = NULL,
  b.lim = NULL,
  l.lim = NULL,
  main = NULL,
  new.mar = NULL,
  bg = NULL,
  fg = NULL,
  col = NULL,
  cex.main = NULL,
  lang = NULL,
  nlevels = 10,
  levels = NA,
  labels = NULL,
  xlim = NA,
  ylim = NA,
  zlim = NA,
  labcex = 1,
  drawlabels = TRUE,
  method = "flattest",
  axes = TRUE,
  frame.plot = NA,
  lty = NA,
  lwd = NA,
  blr.clock = NULL,
  tlr.an = NULL,
  blr.tx = NULL,
  text.sum = NULL,
  base.css.ps.lim = NULL,
  ...
)
}
\arguments{
\item{geo}{See \code{\link[soiltexture]{TT.plot}}}

\item{x}{A data.frame of matrix with 3 columns: x, y and z.}

\item{add}{If TRUE adds the contour to an existing plot.}

\item{tri.sum.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.pos.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.tol}{See \code{\link[soiltexture]{TT.plot}}}

\item{unit.ps}{See \code{\link[soiltexture]{TT.plot}}}

\item{unit.tx}{See \code{\link[soiltexture]{TT.plot}}}

\item{b.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{l.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{main}{See \code{\link[soiltexture]{TT.plot}}}

\item{new.mar}{See \code{\link[soiltexture]{TT.plot}}}

\item{bg}{See \code{\link[soiltexture]{TT.plot}}}

\item{fg}{See \code{\link[soiltexture]{TT.plot}}}

\item{col}{See \code{\link[soiltexture]{TT.plot}}}

\item{cex.main}{See \code{\link[soiltexture]{TT.plot}}}

\item{lang}{See \code{\link[soiltexture]{TT.plot}}}

\item{nlevels}{See \code{\link[graphics]{contour}}}

\item{levels}{See \code{\link[graphics]{contour}}}

\item{labels}{See \code{\link[graphics]{contour}}}

\item{xlim}{See \code{\link[graphics]{contour}}}

\item{ylim}{See \code{\link[graphics]{contour}}}

\item{zlim}{See \code{\link[graphics]{contour}}}

\item{labcex}{See \code{\link[graphics]{contour}}}

\item{drawlabels}{See \code{\link[graphics]{contour}}}

\item{method}{See \code{\link[graphics]{contour}}}

\item{axes}{See \code{\link[graphics]{contour}}}

\item{frame.plot}{See \code{\link[graphics]{contour}}}

\item{lty}{See \code{\link[graphics]{contour}}}

\item{lwd}{See \code{\link[graphics]{contour}}}

\item{blr.clock}{See \code{\link[soiltexture]{TT.plot}}}

\item{tlr.an}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.tx}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.sum}{See \code{\link[soiltexture]{TT.plot}}}

\item{base.css.ps.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{\dots}{Additional parameters passed to  See \code{\link[graphics]{contour}}}
}
\description{
A wrapper for the contour() function adapted to texture triangles (plot
preparation). designed to plot the results of TT.mahalanobis() or
TT.kde2d(), before or after plot.
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
