\name{TT.css2xy}
\alias{TT.css2xy}
\title{TT css2xy}
\description{
}
\usage{
TT.css2xy(tri.data, geo, css.names = NULL, text.tol = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, set.par = FALSE)
}
\arguments{
  \item{tri.data}{
}
  \item{geo}{
}
  \item{css.names}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{set.par}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
