\name{TT.data.test.X}
\alias{TT.data.test.X}
\title{TT data test X}
\description{
}
\usage{
TT.data.test.X(tri.data, text.sum = NULL, text.tol = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL)
}
\arguments{
  \item{tri.data}{
}
  \item{text.sum}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
