\name{TT.deg2rad}
\alias{TT.deg2rad}
\title{Function to convert angle in degree to angle in radian.}
\description{Function to convert angle in degree to angle in radian.}
\usage{
TT.deg2rad(A)
}
\arguments{
  \item{A}{Angle in Degrees}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





