\name{TT.dia2phi}
\alias{TT.dia2phi}
\title{Convert a soil particle diameter dia [micro-meters] into phi = -log2(dia/1000)}
\description{Convert a soil particle diameter dia [micro-meters] into 
phi = -log2(dia). See also TT.phi2dia().}
\usage{TT.dia2phi(dia)}
\arguments{
  \item{dia}{Particle size diameter in micro-meters (will be converted in milli-meters)}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





