\name{TT.DJ.col}
\alias{TT.DJ.col}
\title{A function to obtaine a weight average 'mix' of different colors!}
\description{A function to obtaine a weight average 'mix' of different colors!}
\usage{TT.DJ.col(cl, w, gray.l = FALSE)}
\arguments{
  \item{cl}{
}
  \item{w}{
}
  \item{gray.l}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.}





