\name{TT.edges}
\alias{TT.edges}
\title{TT edges}

\usage{
TT.edges(geo, text.tol = NULL, text.sum = NULL, blr.clock = NULL, col.axis = NULL, plot.axis = TRUE, frame.bg.col = NULL, lwd.axis = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, bg = NULL)
}
\arguments{
  \item{geo}{
}
  \item{text.tol}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
  \item{col.axis}{
}
  \item{plot.axis}{
}
  \item{frame.bg.col}{
}
  \item{lwd.axis}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{bg}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





