\name{TT.gen.op.set}
\alias{TT.gen.op.set}
\title{Internal. Retrieve and set default values from options. }
\description{Retrieve and set default values from options (that do _not_ superseed par()). }
\usage{
TT.gen.op.set(param, assign.op = TRUE, p.env = parent.frame())
}
\arguments{
  \item{param}{
}
  \item{assign.op}{
}
  \item{p.env}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





