% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.geo.get}
\alias{TT.geo.get}
\title{Retrieve and return the geometrical parameters from a list of
parameter values (NULL or not).}
\usage{
TT.geo.get(
  class.sys = NULL,
  blr.clock = NULL,
  tlr.an = NULL,
  blr.tx = NULL,
  text.sum = NULL,
  base.css.ps.lim = NULL
)
}
\arguments{
\item{class.sys}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.clock}{See \code{\link[soiltexture]{TT.plot}}}

\item{tlr.an}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.tx}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.sum}{See \code{\link[soiltexture]{TT.plot}}}

\item{base.css.ps.lim}{See \code{\link[soiltexture]{TT.plot}}}
}
\description{
Retrieve and return the geometrical parameters from a list of parameter
values (NULL or not).
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
