\name{TT.geo.set}
\alias{TT.geo.set}
\title{TT geo set}
\description{
}
\usage{
TT.geo.set(geo, p.env = parent.frame())
}
\arguments{
  \item{geo}{
}
  \item{p.env}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
