\name{TT.get}
\alias{TT.get}
\title{TT get}

\usage{
TT.get(..., par.list = "TT.par", bkp.par.list = "TT.par.bkp", par.env = TT.env)
}
\arguments{
  \item{\dots}{
}
  \item{par.list}{
}
  \item{bkp.par.list}{
}
  \item{par.env}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





