% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.grid}
\alias{TT.grid}
\title{Plot a grid at regular texture intervals inside an existing soil texture
triangle.}
\usage{
TT.grid(
  geo = geo,
  at = NULL,
  text.tol = NULL,
  text.sum = NULL,
  blr.clock = NULL,
  grid.col = NULL,
  grid.lty = NULL,
  lwd.axis = NULL,
  tri.sum.tst = NULL,
  tri.pos.tst = NULL,
  class.p.bg.col = NULL,
  class.p.bg.hue = NULL,
  frame.bg.col = NULL,
  bg = NULL,
  col.axis = NULL
)
}
\arguments{
\item{geo}{See \code{\link[soiltexture]{TT.plot}}}

\item{at}{Vector of numeric values.}

\item{text.tol}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.sum}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.clock}{See \code{\link[soiltexture]{TT.plot}}}

\item{grid.col}{Passed to argument 'col' of \code{\link[graphics]{segments}}}

\item{grid.lty}{Passed to argument 'lty' of \code{\link[graphics]{segments}}}

\item{lwd.axis}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.sum.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.pos.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{class.p.bg.col}{See \code{\link[soiltexture]{TT.plot}}}

\item{class.p.bg.hue}{See \code{\link[soiltexture]{TT.plot}}}

\item{frame.bg.col}{See \code{\link[soiltexture]{TT.plot}}}

\item{bg}{See \code{\link[soiltexture]{TT.plot}}}

\item{col.axis}{See \code{\link[soiltexture]{TT.plot}}}
}
\description{
Plot a grid at regular texture intervals inside an existing soil texture
triangle.
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
