\name{TT.ifelse}
\alias{TT.ifelse}
\title{TT ifelse}
\description{
}
\usage{
TT.ifelse(test, yes, no)
}
\arguments{
  \item{test}{
}
  \item{yes}{
}
  \item{no}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
