% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.iwd}
\alias{TT.iwd}
\title{Inverse weighted distance interpolation on a grid.}
\usage{
TT.iwd(
  tri.data,
  z.name,
  geo,
  css.names = NULL,
  text.tol = NULL,
  text.sum = NULL,
  blr.clock = NULL,
  tri.sum.tst = NULL,
  tri.pos.tst = NULL,
  set.par = FALSE,
  n = 25,
  lims = c("points", "triangle")[1],
  max.dist = NULL,
  q.max.dist = 0.5,
  pow = 0.5
)
}
\arguments{
\item{tri.data}{See \code{\link[soiltexture]{TT.plot}}}

\item{z.name}{See \code{\link[soiltexture]{TT.plot}}}

\item{geo}{See \code{\link[soiltexture]{TT.plot}}}

\item{css.names}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.tol}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.sum}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.clock}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.sum.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.pos.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{set.par}{If TRUE parameters are set automatically to their defualt value}

\item{n}{Number of points in the grid}

\item{lims}{"points" or "triangle"}

\item{max.dist}{Maximum distance}

\item{q.max.dist}{Passed to \code{\link[stats]{quantile}}}

\item{pow}{Single numeric value. Power of the inverse weighted distance}
}
\description{
Inverse weighted distance interpolation on a grid.
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
