\name{TT.kde2d}
\alias{TT.kde2d}
\title{TT kde2d}

\usage{
TT.kde2d(geo, tri.data, css.names = NULL, text.tol = NULL, text.sum = NULL, blr.clock = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, set.par = FALSE, n = 25, lims = c("points", "triangle")[2])
}
\arguments{
  \item{geo}{
}
  \item{tri.data}{
}
  \item{css.names}{
}
  \item{text.tol}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{set.par}{
}
  \item{n}{
}
  \item{lims}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





