\name{TT.lines}
\alias{TT.lines}
\title{TT lines}
\description{
}
\usage{
TT.lines(geo = geo, at.1.s = TT.get("at"), at.2.s = 1 - TT.get("at"), at.3.s = 0, at.1.e = TT.get("at"), at.2.e = 0, at.3.e = 1 - TT.get("at"), text.tol = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL)
}
\arguments{
  \item{geo}{
}
  \item{at.1.s}{
}
  \item{at.2.s}{
}
  \item{at.3.s}{
}
  \item{at.1.e}{
}
  \item{at.2.e}{
}
  \item{at.3.e}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
