\name{TT.locator}
\alias{TT.locator}
\title{Interactive (mouse clic) retrieval the CLAY SILT SAND coordinate of points on a texture triangle.}
\description{Interactive (mouse clic) retrieval the CLAY SILT SAND coordinate of points on a texture triangle. }
\usage{TT.locator(geo, css.names = NULL, text.tol = NULL, 
    tri.sum.tst = NULL, tri.pos.tst = FALSE, set.par = FALSE, 
    n = 512, type = "n", ...)}
\arguments{
  \item{geo}{
}
  \item{css.names}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{set.par}{
}
  \item{n}{
}
  \item{type}{
}
  \item{\dots}{Further argumets passed to locator() }
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





