\name{TT.normalise.sum}
\alias{TT.normalise.sum}
\title{TT normalise sum}
\description{
}
\usage{
TT.normalise.sum(tri.data, css.names = NULL, text.sum = NULL, text.tol = NULL, tri.pos.tst = NULL, residuals = FALSE)
}
\arguments{
  \item{tri.data}{
}
  \item{css.names}{
}
  \item{text.sum}{
}
  \item{text.tol}{
}
  \item{tri.pos.tst}{
}
  \item{residuals}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
