\name{TT.phi2dia}
\alias{TT.phi2dia}
\title{Convert a soil particle phi value into diameter dia [micro-meters]. }
\description{Convert a soil particle phi value into diameter dia [micro-meters]. 
See also TT.dia2phi(). dia = (2^-phi)*1000. Not used by the package. }
\usage{
TT.phi2dia(phi)
}
\arguments{
  \item{phi}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





