\name{TT.points.in.classes}
\alias{TT.points.in.classes}
\title{TT points in classes}

\usage{
TT.points.in.classes(tri.data, class.sys = NULL, PiC.type = NULL, css.names = NULL, text.sum = NULL, base.css.ps.lim = NULL, tri.css.ps.lim = NULL, dat.css.ps.lim = NULL, css.transf = NULL, text.transf.fun = NULL, trsf.add.opt1 = NULL, trsf.add.opt2 = NULL, text.tol = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, collapse = ", ", texture2xy = FALSE, blr.tx = NULL, blr.clock = NULL)
}
\arguments{
  \item{tri.data}{
}
  \item{class.sys}{
}
  \item{PiC.type}{
}
  \item{css.names}{
}
  \item{text.sum}{
}
  \item{base.css.ps.lim}{
}
  \item{tri.css.ps.lim}{
}
  \item{dat.css.ps.lim}{
}
  \item{css.transf}{
}
  \item{text.transf.fun}{
}
  \item{trsf.add.opt1}{
}
  \item{trsf.add.opt2}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{collapse}{
}
  \item{texture2xy}{Single logical. Set to FALSE to avoid any transformation of the 
texture data (trigonometric) prior to testure data classification. 
Setting to FALSE avoid some numerical accuracy problems when 
a point is on the border of a texture class.}
  \item{blr.tx}{
}
  \item{blr.clock}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





