\name{TT.str}
\alias{TT.str}
\title{TT str}
\description{
}
\usage{
TT.str(x, str.min = 0, str.max = 1)
}
\arguments{
  \item{x}{
}
  \item{str.min}{
}
  \item{str.max}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
