\name{TT.switch}
\alias{TT.switch}
\title{TT switch}
\description{
}
\usage{
TT.switch(blr.clock = TT.get("blr.clock"), c1 = NA, c2 = NA, c3 = NA, c4 = NA, blr.order = c(1, 3, 2))
}
\arguments{
  \item{blr.clock}{
}
  \item{c1}{
}
  \item{c2}{
}
  \item{c3}{
}
  \item{c4}{
}
  \item{blr.order}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
