% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.text}
\alias{TT.text}
\title{Plot text labels for each values of a soil texture data table on an existing
texture plot.}
\usage{
TT.text(
  tri.data,
  geo,
  labels = NULL,
  css.names = NULL,
  base.css.ps.lim = NULL,
  dat.css.ps.lim = NULL,
  css.transf = NULL,
  text.transf.fun = NULL,
  trsf.add.opt1 = NULL,
  trsf.add.opt2 = NULL,
  text.tol = NULL,
  text.sum = NULL,
  blr.clock = NULL,
  fg = NULL,
  col = NULL,
  cex = NULL,
  font = NULL,
  family.op = NULL,
  adj = NULL,
  pos = NULL,
  offset = NULL,
  tri.sum.tst = NULL,
  tri.pos.tst = NULL,
  blr.tx = NULL
)
}
\arguments{
\item{tri.data}{See \code{\link[soiltexture]{TT.plot}}}

\item{geo}{See \code{\link[soiltexture]{TT.plot}}}

\item{labels}{Vector of character strings. Label to be plotted, for each point in 'tri.data'}

\item{css.names}{See \code{\link[soiltexture]{TT.plot}}}

\item{base.css.ps.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{dat.css.ps.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{css.transf}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.transf.fun}{See \code{\link[soiltexture]{TT.plot}}}

\item{trsf.add.opt1}{See \code{\link[soiltexture]{TT.plot}}}

\item{trsf.add.opt2}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.tol}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.sum}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.clock}{See \code{\link[soiltexture]{TT.plot}}}

\item{fg}{See \code{\link[soiltexture]{TT.plot}}}

\item{col}{See \code{\link[graphics]{text}}}

\item{cex}{See \code{\link[graphics]{text}}}

\item{font}{See \code{\link[graphics]{text}}}

\item{family.op}{See \code{\link[soiltexture]{TT.plot}}}

\item{adj}{See \code{\link[graphics]{text}}}

\item{pos}{See \code{\link[graphics]{text}}}

\item{offset}{See \code{\link[graphics]{text}}}

\item{tri.sum.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.pos.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.tx}{See \code{\link[soiltexture]{TT.plot}}}
}
\description{
Plot text labels for each values of a soil texture data table on an existing
texture plot.
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
