% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.text.transf}
\alias{TT.text.transf}
\title{Log-linear transformation of a soil texture data table between 2 particle
size systems (3 classes).}
\usage{
TT.text.transf(
  tri.data,
  base.css.ps.lim,
  dat.css.ps.lim,
  css.names = NULL,
  blr.tx = NULL,
  text.sum = NULL,
  text.tol = NULL,
  tri.sum.tst = NULL,
  tri.pos.tst = NULL,
  trsf.add.opt1 = NULL,
  trsf.add.opt2 = NULL
)
}
\arguments{
\item{tri.data}{See \code{\link[soiltexture]{TT.plot}}}

\item{base.css.ps.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{dat.css.ps.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{css.names}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.tx}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.sum}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.tol}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.sum.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.pos.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{trsf.add.opt1}{See \code{\link[soiltexture]{TT.plot}}}

\item{trsf.add.opt2}{See \code{\link[soiltexture]{TT.plot}}}
}
\description{
Log-linear transformation of a soil texture data table ('tri.data') from one
particle size system ('dat.css.ps.lim') into another ('base.css.ps.lim').
Only 3 particle size classes allowed. See TT.text.transf.X for
transformation involving more than 3 particle classes. 'tri.data' may
contain other variables (not in 'css.names'). They are returned unchanged
with the transformed texture data.
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
