\name{TT.text.transf.X}
\alias{TT.text.transf.X}
\title{TT text transf X}
\description{
}
\usage{
TT.text.transf.X(tri.data, base.ps.lim, dat.ps.lim, text.sum = NULL, text.tol = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL)
}
\arguments{
  \item{tri.data}{
}
  \item{base.ps.lim}{
}
  \item{dat.ps.lim}{
}
  \item{text.sum}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
