\name{TT.vertices.tbl}
\alias{TT.vertices.tbl}
\title{TT vertices tbl}
\description{
}
\usage{
TT.vertices.tbl(class.sys = "FAO50.TT")
}
\arguments{
  \item{class.sys}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
