\name{TT.xy2css}
\alias{TT.xy2css}
\title{Internal. Convert point-data duplets (2 variables, x-y coordinaes) in Clay silta and sand coordinates}

\usage{
TT.xy2css(xy.data, geo, css.names = NULL, text.tol = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, set.par = FALSE, blr.clock = NULL, text.sum = NULL)
}
\arguments{
  \item{xy.data}{
}
  \item{geo}{
}
  \item{css.names}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{set.par}{
}
  \item{blr.clock}{
}
  \item{text.sum}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





