\name{auto.set}
\alias{auto.set}
\title{auto set}

\usage{
auto.set(fun = sys.function(which = -1), assign.op = TRUE, p.env = parent.frame(), set.par = TRUE)
}
\arguments{
  \item{fun}{
}
  \item{assign.op}{
}
  \item{p.env}{
}
  \item{set.par}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





