\name{soiltexture-package}
\alias{soiltexture-package}
\alias{soiltexture}
\docType{package}
\title{
\packageTitle{soiltexture}
}
\description{
\packageDescription{soiltexture}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{soiltexture}
\packageIndices{soiltexture}
}
\author{
\packageAuthor{soiltexture}

Maintainer: \packageMaintainer{soiltexture}
}
\keyword{ package }
