\name{soiltexture-package}
\alias{soiltexture-package}
\alias{soiltexture}
\docType{package}
\title{Functions for soil texture plot, classification and transformation}
\description{"The Soil Texture Wizard" is a set of R functions designed to produce
texture triangles (also called texture plots, texture diagrams, texture ternary plots),
classify and transform soil textures data. These functions virtually allows to plot any
soil texture triangle / classification into any triangle geometry (isosceles,
right-angled triangles, etc.). This set of function is expected to be useful to people
using soil textures data from different soil texture classification or different particle
size systems. Several texture triangles are predefined: USDA; FAO (which is also the
triangle for the soil map of Europe); Aisne (France); GEPPA (France); German triangles
"Bodenkundliche Kartieranleitung 1994", "SEA 1974" and "TGL 1985"; Soil
Survey of England and Wales; Australian triangle; Belgian triangle; Canadian triangle;
ISSS triangle; Romanian triangle; Polish triangle.}
\details{
\tabular{ll}{Package: \tab soiltexture\cr
Version: \tab 1.2.13\cr
Date: \tab 2014-01-10\cr
Title: \tab Functions for soil texture plot, classification and transformation\cr
Author: \tab Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.\cr
Maintainer: \tab Julien MOEYS <jules_m78-soiltexture@yahoo.fr>\cr
Depends: \tab R (>= 3.0.2), sp, MASS\cr
Suggests: \tab xtable\cr
License: \tab AGPL (>=3)\cr
URL: \tab http://soiltexture.r-forge.r-project.org/\cr}
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.}

\keyword{ package }


