\name{soiltexture-package}
\alias{soiltexture-package}
\docType{package}
\title{Functions for soil texture plot, classification and transformation}
\description{"The Soil Texture Wizard" is a set of R functions designed to produce
texture triangles (also called texture plots, texture diagrams, texture ternary plots),
classify and transform soil textures data. These functions virtually allows to plot any
soil texture triangle / classification into any triangle geometry (isosceles,
right-angled triangles, etc.). This set of function is expected to be useful to people
using soil textures data from different soil texture classification or different particle
size systems. Several texture triangles are predefined: USDA; FAO (which is also the
triangle for the soil map of Europe); Aisne (France); GEPPA (France); German triangle
(Bodenkundliche Kartieranleitung 1994); Soil Survey of England and Wales (UK); Australian
triangle; Belgian triangle; Canadian triangle; ISSS triangle; Romanian traingle. The Soil Texture Wizard is initially
develloped from the functions 'soil.texture()' from the package 'plotrix' by Jim Lemon
et al.}
\details{
\tabular{ll}{Package: \tab soiltexture\cr
Version: \tab 1.0\cr
Date: \tab 2010-08-30\cr
Title: \tab Functions for soil texture plot, classification and transformation\cr
Author: \tab Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.\cr
Maintainer: \tab Julien MOEYS <jules_m78-soiltexture@yahoo.fr>\cr
Depends: \tab R (>= 2.11.1), sp, MASS, drc\cr
Suggests: \tab plotrix\cr
License: \tab GPL (>=3)\cr
URL: \tab http://soiltexture.r-forge.r-project.org/\cr}
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}

\keyword{ package }


