% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture-package.R
\docType{package}
\name{soiltexture-package}
\alias{soiltexture}
\alias{soiltexture-package}
\title{soiltexture: Functions for Soil Texture Plot, Classification and Transformation}
\description{
"The Soil Texture Wizard" is a set of R functions designed to produce texture triangles (also called texture plots, texture diagrams, texture ternary plots), classify and transform soil textures data. These functions virtually allows to plot any soil texture triangle (classification) into any triangle geometry (isosceles, right-angled triangles, etc.). This set of function is expected to be useful to people using soil textures data from different soil texture classification or different particle size systems. Many (> 15) texture triangles from all around the world are predefined in the package. A simple text based graphical user interface is provided: soiltexture_gui().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/julienmoeys/soiltexture}
  \item Report bugs at \url{https://github.com/julienmoeys/soiltexture/issues}
}

}
\author{
\strong{Maintainer}: Julien Moeys \email{jules_m78-soiltexture@yahoo.fr}

Other contributors:
\itemize{
  \item Wei Shangguan [contributor]
  \item Rainer Petzold [contributor]
  \item Budiman Minasny [contributor]
  \item Bogdan Rosca [contributor]
  \item Nic Jelinski [contributor]
  \item Wiktor Zelazny [contributor]
  \item Rodolfo Marcondes Silva Souza [contributor]
  \item Jose Lucas Safanelli [contributor]
  \item Alexandre ten Caten [contributor]
}

}
\keyword{internal}
