\name{soiltexture-package}
\alias{soiltexture-package}
\alias{soiltexture}
\docType{package}
\title{Functions for Soil Texture Plot, Classification and Transformation}
\description{"The Soil Texture Wizard" is a set of R functions designed to produce texture triangles (also called texture plots, texture diagrams, texture ternary plots), classify and transform soil textures data. These functions virtually allows to plot any soil texture triangle (classification) into any triangle geometry (isosceles, right-angled triangles, etc.). This set of function is expected to be useful to people using soil textures data from different soil texture classification or different particle size systems. Many (> 15) texture triangles from all around the world are predefined in the package. A simple text based graphical user interface is provided: soiltexture_gui().}
\details{
\tabular{ll}{Package: \tab soiltexture\cr
Version: \tab 1.3.3\cr
Date: \tab 2015-05-15\cr
Title: \tab Functions for Soil Texture Plot, Classification and Transformation\cr
Author: \tab Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]\cr
Authors@R: \tab c( person( "Julien", "Moeys", role = c("aut", "cre"), email = "jules_m78-soiltexture@yahoo.fr" ), person( "Wei", "Shangguan", role = "ctb" ), person( "Rainer", "Petzold", role = "ctb" ), person( "Budiman", "Minasny", role = "ctb" ), person( "Bogdan", "Rosca", role = "ctb" ), person( "Nic", "Jelinski", role = "ctb" ), person( "Wiktor", "Zelazny", role = "ctb" ), person( "Rodolfo", "Marcondes Silva Souza", role = "ctb" ), person( "Jose Lucas", "Safanelli", role = "ctb" ), person( "Alexandre", "ten Caten", role = "ctb" ) )\cr
Maintainer: \tab Julien Moeys <jules_m78-soiltexture@yahoo.fr>\cr
Depends: \tab R (>= 3.1.1), utils\cr
Suggests: \tab xtable\cr
License: \tab AGPL (>=3)\cr
URL: \tab http://soiltexture.r-forge.r-project.org/\cr
Imports: \tab sp, MASS, tools, tcltk\cr}
}
\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}

\keyword{ package }


