\name{TT.add}
\alias{TT.add}
\title{TT add}

\usage{
TT.add(..., par.list = "TT.par", bkp.par.list = "TT.par.bkp", par.env = TT.env)
}
\arguments{
  \item{\dots}{
}
  \item{par.list}{
}
  \item{bkp.par.list}{
}
  \item{par.env}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





