\name{TT.add}
\alias{TT.add}
\title{Function to add a new default package parameters. }
\description{Function to add a new default package parameters. Mostly used 
to add a new texture triangle definition.}
\usage{TT.add(..., par.list = "TT.par", bkp.par.list = "TT.par.bkp", 
    par.env = TT.env)}
\arguments{
  \item{\dots}{
}
  \item{par.list}{
}
  \item{bkp.par.list}{
}
  \item{par.env}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





