\name{TT.auto.set}
\alias{TT.auto.set}
\title{Internal. Retrieve and set default values for parameters (par() or not), when NULL.}
\description{Retrieve and set default values for parameters (par() or not), when NULL.}
\usage{
TT.auto.set(fun = sys.function(which = -1),
assign.op = TRUE,
p.env = parent.frame(),
set.par = TRUE)
}
\arguments{
  \item{fun}{
}
  \item{assign.op}{
}
  \item{p.env}{
}
  \item{set.par}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.}





