% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.baseplot}
\alias{TT.baseplot}
\title{Internal. Create an empty plot scene for a texture triangle.}
\usage{
TT.baseplot(
  geo = NULL,
  class.sys = "none",
  blr.clock = NULL,
  tlr.an = NULL,
  blr.tx = NULL,
  text.sum = NULL,
  base.css.ps.lim = NULL,
  tri.sum.tst = NULL,
  tri.pos.tst = NULL,
  text.tol = NULL,
  unit.ps = NULL,
  unit.tx = NULL,
  b.lim = NULL,
  l.lim = NULL,
  main = NULL,
  new.mar = NULL,
  bg = NULL,
  fg = NULL,
  col = NULL,
  cex.main = NULL,
  lang = NULL
)
}
\arguments{
\item{geo}{See \code{\link[soiltexture]{TT.plot}}}

\item{class.sys}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.clock}{See \code{\link[soiltexture]{TT.plot}}}

\item{tlr.an}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.tx}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.sum}{See \code{\link[soiltexture]{TT.plot}}}

\item{base.css.ps.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.sum.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.pos.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.tol}{See \code{\link[soiltexture]{TT.plot}}}

\item{unit.ps}{See \code{\link[soiltexture]{TT.plot}}}

\item{unit.tx}{See \code{\link[soiltexture]{TT.plot}}}

\item{b.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{l.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{main}{See \code{\link[soiltexture]{TT.plot}}}

\item{new.mar}{See \code{\link[soiltexture]{TT.plot}}}

\item{bg}{See \code{\link[soiltexture]{TT.plot}}}

\item{fg}{See \code{\link[soiltexture]{TT.plot}}}

\item{col}{See \code{\link[soiltexture]{TT.plot}}}

\item{cex.main}{See \code{\link[soiltexture]{TT.plot}}}

\item{lang}{See \code{\link[soiltexture]{TT.plot}}}
}
\description{
Create an empty plot where a texture triangle can be drawn with other
secondary functions (frame, axis, ...). Also return the 'geo' parameters
needed by these secondary functions.
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
