\name{TT.blr.ps.lim}
\alias{TT.blr.ps.lim}
\title{Internal. Create a tabular version of clay silt sand particle size limits. }
\description{Create a tabular version of clay silt sand particle size limits. }
\usage{TT.blr.ps.lim(blr.tx, css.ps.lim)}
\arguments{
  \item{blr.tx}{
}
  \item{css.ps.lim}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





