\name{TT.blr.ps.lim}
\alias{TT.blr.ps.lim}
\title{Internal. Create a tabular version of clay silt sand particle size limits. }
\description{Create a tabular version of clay silt sand particle size limits. }
\usage{
TT.blr.ps.lim(blr.tx, css.ps.lim)
}
\arguments{
  \item{blr.tx}{
}
  \item{css.ps.lim}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





