\name{TT.blr.tx.check}
\alias{TT.blr.tx.check}
\title{Internal. Check the consistency between blr.tx and css.names. }
\description{Check the consistency between blr.tx and css.names. All values 
in blr.tx should be found in css.names and vice-versa.}
\usage{TT.blr.tx.check(blr.tx, css.names)}
\arguments{
  \item{blr.tx}{
}
  \item{css.names}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





