\name{TT.check.ps.lim}
\alias{TT.check.ps.lim}
\title{TT check ps lim}
\description{
}
\usage{
TT.check.ps.lim(base.ps.lim, dat.ps.lim, ps.lim.length = c(4, 4))
}
\arguments{
  \item{base.ps.lim}{
}
  \item{dat.ps.lim}{
}
  \item{ps.lim.length}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
