\name{TT.chemometrics.alr}
\alias{TT.chemometrics.alr}
\title{TT chemometrics alr}
\description{
}
\usage{
TT.chemometrics.alr(X, divisorvar, css.names)
}
\arguments{
  \item{X}{
}
  \item{divisorvar}{
}
  \item{css.names}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
