% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.classes}
\alias{TT.classes}
\title{Plot the texture classes polygons in a texture triangle plot.}
\usage{
TT.classes(
  geo,
  class.sys,
  tri.css.ps.lim = NULL,
  css.transf = NULL,
  text.transf.fun = NULL,
  trsf.add.opt1 = NULL,
  trsf.add.opt2 = NULL,
  text.tol = NULL,
  text.sum = NULL,
  base.css.ps.lim = NULL,
  blr.tx = NULL,
  blr.clock = NULL,
  tri.sum.tst = NULL,
  tri.pos.tst = NULL,
  bg = NULL,
  class.lab.col = NULL,
  class.p.bg.col = NULL,
  class.p.bg.hue = NULL,
  class.line.col = NULL,
  class.lty = NULL,
  class.lab.show = NULL,
  cex.lab = NULL,
  font.lab = NULL,
  family.op = NULL,
  lwd.axis = NULL,
  col.axis = NULL,
  new.centroid = TRUE
)
}
\arguments{
\item{geo}{See \code{\link[soiltexture]{TT.plot}}}

\item{class.sys}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.css.ps.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{css.transf}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.transf.fun}{See \code{\link[soiltexture]{TT.plot}}}

\item{trsf.add.opt1}{See \code{\link[soiltexture]{TT.plot}}}

\item{trsf.add.opt2}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.tol}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.sum}{See \code{\link[soiltexture]{TT.plot}}}

\item{base.css.ps.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.tx}{See See \code{\link[soiltexture]{TT.plot}}}

\item{blr.clock}{See See \code{\link[soiltexture]{TT.plot}}}

\item{tri.sum.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.pos.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{bg}{See \code{\link[soiltexture]{TT.plot}}}

\item{class.lab.col}{See \code{\link[soiltexture]{TT.plot}}}

\item{class.p.bg.col}{See \code{\link[soiltexture]{TT.plot}}}

\item{class.p.bg.hue}{See \code{\link[soiltexture]{TT.plot}}}

\item{class.line.col}{See \code{\link[soiltexture]{TT.plot}}}

\item{class.lty}{See \code{\link[soiltexture]{TT.plot}}}

\item{class.lab.show}{See \code{\link[soiltexture]{TT.plot}}}

\item{cex.lab}{See \code{\link[soiltexture]{TT.plot}}}

\item{font.lab}{See \code{\link[soiltexture]{TT.plot}}}

\item{family.op}{See \code{\link[soiltexture]{TT.plot}}}

\item{lwd.axis}{See \code{\link[soiltexture]{TT.plot}}}

\item{col.axis}{See \code{\link[soiltexture]{TT.plot}}}

\item{new.centroid}{Single logical. If TRUE (default) the new method (Paul
Bourke) is used to calculate the centroid. If FALSE the centroid is taken as
the mean x and y coordinates of the vertices.}
}
\description{
Plot the texture classes ploygons in an existing texture triangle plot. Draw
the polygons and the labels inside each polygons.
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
