\name{TT.classes}
\alias{TT.classes}
\title{TT classes}
\description{
}
\usage{
TT.classes(geo, class.sys, tri.css.ps.lim = NULL, css.transf = NULL, text.transf.fun = NULL, trsf.add.opt1 = NULL, trsf.add.opt2 = NULL, text.tol = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, bg = NULL, class.lab.col = NULL, class.p.bg.col = NULL, class.p.bg.hue = NULL, class.line.col = NULL, class.lty = NULL, class.lab.show = NULL, cex.lab = NULL, font.lab = NULL, family.op = NULL, lwd.axis = NULL, col.axis = NULL)
}
\arguments{
  \item{geo}{
}
  \item{class.sys}{
}
  \item{tri.css.ps.lim}{
}
  \item{css.transf}{
}
  \item{text.transf.fun}{
}
  \item{trsf.add.opt1}{
}
  \item{trsf.add.opt2}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{bg}{
}
  \item{class.lab.col}{
}
  \item{class.p.bg.col}{
}
  \item{class.p.bg.hue}{
}
  \item{class.line.col}{
}
  \item{class.lty}{
}
  \item{class.lab.show}{
}
  \item{cex.lab}{
}
  \item{font.lab}{
}
  \item{family.op}{
}
  \item{lwd.axis}{
}
  \item{col.axis}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
