\name{TT.classes}
\alias{TT.classes}
\title{Plot the texture classes polygons in a texture triangle plot.}
\description{Plot the texture classes ploygons in an existing texture 
triangle plot. Draw the polygons and the labels inside each 
polygons.}
\usage{
TT.classes(geo,
class.sys,
tri.css.ps.lim = NULL,
css.transf = NULL,
text.transf.fun = NULL,
trsf.add.opt1 = NULL,
trsf.add.opt2 = NULL,
text.tol = NULL,
text.sum = NULL,
base.css.ps.lim = NULL,
blr.tx = NULL,
blr.clock = NULL,
tri.sum.tst = NULL,
tri.pos.tst = NULL,
bg = NULL,
class.lab.col = NULL,
class.p.bg.col = NULL,
class.p.bg.hue = NULL,
class.line.col = NULL,
class.lty = NULL,
class.lab.show = NULL,
cex.lab = NULL,
font.lab = NULL,
family.op = NULL,
lwd.axis = NULL,
col.axis = NULL,
new.centroid = TRUE)
}
\arguments{
  \item{geo}{
}
  \item{class.sys}{
}
  \item{tri.css.ps.lim}{
}
  \item{css.transf}{
}
  \item{text.transf.fun}{
}
  \item{trsf.add.opt1}{
}
  \item{trsf.add.opt2}{
}
  \item{text.tol}{
}
  \item{text.sum}{
}
  \item{base.css.ps.lim}{
}
  \item{blr.tx}{
}
  \item{blr.clock}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{bg}{
}
  \item{class.lab.col}{
}
  \item{class.p.bg.col}{
}
  \item{class.p.bg.hue}{
}
  \item{class.line.col}{
}
  \item{class.lty}{
}
  \item{class.lab.show}{
}
  \item{cex.lab}{
}
  \item{font.lab}{
}
  \item{family.op}{
}
  \item{lwd.axis}{
}
  \item{col.axis}{
}
  \item{new.centroid}{Single logical. If TRUE (default) the new method (Paul Bourke) 
is used to calculate the centroid. If FALSE the centroid is 
taken as the mean x and y coordinates of the vertices.}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.}





