\name{TT.classes.tbl}
\alias{TT.classes.tbl}
\title{Returns the table of classes of a texture classification system.}
\description{Returns the table of classes of a texture classification system. 
Returns the classes abbreviations, names and the vertices numbers 
that defines each class. Use TT.vertices.tbl() to retrieve the 
clay silt sand coordinates of the triangle classes vertices. 
See also TT.vertices.plot().}
\usage{TT.classes.tbl(class.sys = "HYPRES.TT", collapse = NULL)}
\arguments{
  \item{class.sys}{
}
  \item{collapse}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





