\name{TT.classes.tbl}
\alias{TT.classes.tbl}
\title{Returns the table of classes of a texture classification system.}
\description{Returns the table of classes of a texture classification system. 
Returns the classes abbreviations, names and the vertices numbers 
that defines each class. Use TT.vertices.tbl() to retrieve the 
clay silt sand coordinates of the triangle classes vertices. 
See also TT.vertices.plot().}
\usage{TT.classes.tbl(class.sys = "HYPRES.TT", collapse = NULL)}
\arguments{
  \item{class.sys}{
}
  \item{collapse}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.}





