\name{TT.col2hsv}
\alias{TT.col2hsv}
\title{Internal. Convert any colors to hsv. }
\description{Convert any colors to hsv. Wrapper around rgb2hsv() and col2rgb(). }
\usage{TT.col2hsv(col)}
\arguments{
  \item{col}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





