\name{TT.contour}
\alias{TT.contour}
\title{Wrapper for the contour() function adapted to texture triangles.}
\description{A wrapper for the contour() function 
adapted to texture triangles (plot preparation).
designed to plot the results of TT.mahalanobis() or 
TT.kde2d(), before or after plot.}
\usage{TT.contour(geo, x, add = FALSE, tri.sum.tst = NULL, tri.pos.tst = NULL, 
    text.tol = NULL, unit.ps = NULL, unit.tx = NULL, b.lim = NULL, 
    l.lim = NULL, main = NULL, new.mar = NULL, bg = NULL, fg = NULL, 
    col = NULL, cex.main = NULL, lang = NULL, nlevels = 10, levels = NA, 
    labels = NULL, xlim = NA, ylim = NA, zlim = NA, labcex = 1, 
    drawlabels = TRUE, method = "flattest", axes = TRUE, frame.plot = NA, 
    lty = NA, lwd = NA, blr.clock = NULL, tlr.an = NULL, blr.tx = NULL, 
    text.sum = NULL, base.css.ps.lim = NULL, ...)}
\arguments{
  \item{geo}{
}
  \item{x}{
}
  \item{add}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{text.tol}{
}
  \item{unit.ps}{
}
  \item{unit.tx}{
}
  \item{b.lim}{
}
  \item{l.lim}{
}
  \item{main}{
}
  \item{new.mar}{
}
  \item{bg}{
}
  \item{fg}{
}
  \item{col}{
}
  \item{cex.main}{
}
  \item{lang}{
}
  \item{nlevels}{
}
  \item{levels}{
}
  \item{labels}{
}
  \item{xlim}{
}
  \item{ylim}{
}
  \item{zlim}{
}
  \item{labcex}{
}
  \item{drawlabels}{
}
  \item{method}{
}
  \item{axes}{
}
  \item{frame.plot}{
}
  \item{lty}{
}
  \item{lwd}{
}
  \item{blr.clock}{
}
  \item{tlr.an}{
}
  \item{blr.tx}{
}
  \item{text.sum}{
}
  \item{base.css.ps.lim}{
}
  \item{\dots}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.}





