\name{TT.css2xy}
\alias{TT.css2xy}
\title{TT css2xy}

\usage{
TT.css2xy(tri.data, geo, css.names = NULL, text.tol = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, set.par = FALSE, text.sum = NULL, blr.clock = NULL)
}
\arguments{
  \item{tri.data}{
}
  \item{geo}{
}
  \item{css.names}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{set.par}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





