\name{TT.css2xy}
\alias{TT.css2xy}
\title{Internal. Converts texture data (3 classes) into x-y coordinates. }
\description{Converts texture data (3 classes) into x-y coordinates. This 
function is the 'heart' of most soiltexture plot functions.}
\usage{TT.css2xy(tri.data, geo, css.names = NULL, text.tol = NULL, 
    tri.sum.tst = NULL, tri.pos.tst = NULL, set.par = FALSE, 
    text.sum = NULL, blr.clock = NULL)}
\arguments{
  \item{tri.data}{
}
  \item{geo}{
}
  \item{css.names}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{set.par}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





