\name{TT.data.test}
\alias{TT.data.test}
\title{TT data test}

\usage{
TT.data.test(tri.data, css.names = NULL, text.sum = NULL, text.tol = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL)
}
\arguments{
  \item{tri.data}{
}
  \item{css.names}{
}
  \item{text.sum}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





