% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.dataset}
\alias{TT.dataset}
\title{Genetates a virtual cross correlated clay silt sand + Z dataset.}
\usage{
TT.dataset(n, seed.val = NULL, css.names = NULL, text.sum = NULL)
}
\arguments{
\item{n}{Single numeric value. Number of points to be generated.}

\item{seed.val}{See \code{\link[base]{set.seed}}}

\item{css.names}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.sum}{See \code{\link[soiltexture]{TT.plot}}}
}
\description{
Genetates a virtual cross correlated clay silt sand + Z dataset, where Z is
a virtual 4th variable correlated to the texture.
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
