\name{TT.dataset}
\alias{TT.dataset}
\title{Genetates a virtual cross correlated clay silt sand + Z dataset. }
\description{Genetates a virtual cross correlated clay silt sand + Z dataset, 
where Z is a virtual 4th variable correlated to the texture.}
\usage{
TT.dataset(n, seed.val = NULL, css.names = NULL, text.sum = NULL)
}
\arguments{
  \item{n}{
}
  \item{seed.val}{
}
  \item{css.names}{
}
  \item{text.sum}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





