\name{TT.deg2rad}
\alias{TT.deg2rad}
\title{Internal. Function to convert angle in degree to angle in radian.}
\description{Function to convert angle in degree to angle in radian.}
\usage{TT.deg2rad(A)}
\arguments{
  \item{A}{Angle in Degrees}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





