% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.dia2phi}
\alias{TT.dia2phi}
\title{Convert a soil particle diameter dia [micro-meters] into phi =
-log2(dia/1000)}
\usage{
TT.dia2phi(dia)
}
\arguments{
\item{dia}{Particle size diameter in micro-meters (will be converted in
milli-meters)}
}
\description{
Convert a soil particle diameter dia [micro-meters] into phi = -log2(dia).
See also TT.phi2dia().
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
