\name{TT.DJ.col}
\alias{TT.DJ.col}
\title{Internal. A function to obtaine a weight average 'mix' of different colors!}
\description{A function to obtaine a weight average 'mix' of different colors!}
\usage{TT.DJ.col(cl, w, gray.l = FALSE)}
\arguments{
  \item{cl}{
}
  \item{w}{
}
  \item{gray.l}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





