\name{TT.edges}
\alias{TT.edges}
\title{TT edges}
\description{
}
\usage{
TT.edges(geo, text.tol = NULL, col.axis = NULL, plot.axis = TRUE, frame.bg.col = NULL, lwd.axis = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, bg = NULL)
}
\arguments{
  \item{geo}{
}
  \item{text.tol}{
}
  \item{col.axis}{
}
  \item{plot.axis}{
}
  \item{frame.bg.col}{
}
  \item{lwd.axis}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{bg}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
