\name{TT.edges}
\alias{TT.edges}
\title{Internal. Plot the edges (bare axis) of a soil texture triangle. }
\description{Plot the edges (bare axis) of a soil texture triangle. This 
is not a primary plot function, TT.baseplot() must have been 
called before (usually inside TT.plot()).}
\usage{
TT.edges(geo,
text.tol = NULL,
text.sum = NULL,
blr.clock = NULL,
col.axis = NULL,
plot.axis = TRUE,
frame.bg.col = NULL,
lwd.axis = NULL,
tri.sum.tst = NULL,
tri.pos.tst = NULL,
bg = NULL)
}
\arguments{
  \item{geo}{
}
  \item{text.tol}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
  \item{col.axis}{
}
  \item{plot.axis}{
}
  \item{frame.bg.col}{
}
  \item{lwd.axis}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{bg}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.}





