\name{TT.gen.op.set}
\alias{TT.gen.op.set}
\title{Internal. Retrieve and set default values from options. }
\description{Retrieve and set default values from options (that do _not_ superseed par()). }
\usage{TT.gen.op.set(param, assign.op = TRUE, p.env = parent.frame())}
\arguments{
  \item{param}{
}
  \item{assign.op}{
}
  \item{p.env}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





