\name{TT.geo.get}
\alias{TT.geo.get}
\title{Internal. Retrieve and return the geometrical parameters from a list of parameter values (NULL or not).}
\description{Retrieve and return the geometrical parameters from a list of parameter values (NULL or not).}
\usage{TT.geo.get(class.sys = NULL, blr.clock = NULL, tlr.an = NULL, 
    blr.tx = NULL, text.sum = NULL, base.css.ps.lim = NULL)}
\arguments{
  \item{class.sys}{
}
  \item{blr.clock}{
}
  \item{tlr.an}{
}
  \item{blr.tx}{
}
  \item{text.sum}{
}
  \item{base.css.ps.lim}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





