\name{TT.geo.get}
\alias{TT.geo.get}
\title{TT geo get}

\usage{
TT.geo.get(class.sys = NULL, blr.clock = NULL, tlr.an = NULL, blr.tx = NULL, text.sum = NULL, base.css.ps.lim = NULL)
}
\arguments{
  \item{class.sys}{
}
  \item{blr.clock}{
}
  \item{tlr.an}{
}
  \item{blr.tx}{
}
  \item{text.sum}{
}
  \item{base.css.ps.lim}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





