\name{TT.geo.set}
\alias{TT.geo.set}
\title{Internal. Takes "geo" values and assign them individually in the parent function. }
\description{Takes "geo" values and assign them individually in the parent function. }
\usage{TT.geo.set(geo, p.env = parent.frame())}
\arguments{
  \item{geo}{
}
  \item{p.env}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





