\name{TT.geo.set}
\alias{TT.geo.set}
\title{Internal. Takes "geo" values and assign them individually in the parent function. }
\description{Takes "geo" values and assign them individually in the parent function. }
\usage{
TT.geo.set(geo, p.env = parent.frame())
}
\arguments{
  \item{geo}{
}
  \item{p.env}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





