% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.get}
\alias{TT.get}
\title{Function to retrieve / get the default package parameters.}
\usage{
TT.get(..., par.list = "TT.par", bkp.par.list = "TT.par.bkp", par.env = TT.env)
}
\arguments{
\item{\dots}{Vector of character strings. Name of arguments for which default value is to be retrieved.}

\item{par.list}{Name of the list containing the parameters}

\item{bkp.par.list}{Name of the backuped list containing the default parameters}

\item{par.env}{name of the environment containing the parameter lists}
}
\description{
Function to retrieve / get the default package parameters.
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
