\name{TT.grid}
\alias{TT.grid}
\title{Plot a grid at regular texture intervals inside an existing soil texture triangle. }
\description{Plot a grid at regular texture intervals inside an existing soil texture triangle.}
\usage{TT.grid(geo = geo, at = NULL, text.tol = NULL, text.sum = NULL, 
    blr.clock = NULL, grid.col = NULL, grid.lty = NULL, 
    lwd.axis = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, 
    class.p.bg.col = NULL, class.p.bg.hue = NULL, frame.bg.col = NULL, 
    bg = NULL, col.axis = NULL)}
\arguments{
  \item{geo}{
}
  \item{at}{
}
  \item{text.tol}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
  \item{grid.col}{
}
  \item{grid.lty}{
}
  \item{lwd.axis}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{class.p.bg.col}{
}
  \item{class.p.bg.hue}{
}
  \item{frame.bg.col}{
}
  \item{bg}{
}
  \item{col.axis}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





