\name{TT.ifelse}
\alias{TT.ifelse}
\title{Internal. Flexible version of ifelse. }
\description{Flexible version of ifelse. }
\usage{
TT.ifelse(test, yes, no)
}
\arguments{
  \item{test}{
}
  \item{yes}{
}
  \item{no}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





