\name{TT.ifelse}
\alias{TT.ifelse}
\title{Internal. Flexible version of ifelse. }
\description{Flexible version of ifelse. }
\usage{TT.ifelse(test, yes, no)}
\arguments{
  \item{test}{
}
  \item{yes}{
}
  \item{no}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





