% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.image}
\alias{TT.image}
\title{Wrapper for the image() function adapted to texture triangles.}
\usage{
TT.image(
  geo,
  x,
  add = FALSE,
  tri.sum.tst = NULL,
  tri.pos.tst = NULL,
  text.tol = NULL,
  unit.ps = NULL,
  unit.tx = NULL,
  b.lim = NULL,
  l.lim = NULL,
  main = NULL,
  new.mar = NULL,
  bg = NULL,
  fg = NULL,
  cex.main = NULL,
  lang = NULL,
  xlim = NA,
  ylim = NA,
  zlim = NA,
  col = rev(heat.colors(12)),
  oldstyle = FALSE,
  blr.clock = NULL,
  tlr.an = NULL,
  blr.tx = NULL,
  text.sum = NULL,
  base.css.ps.lim = NULL,
  ...
)
}
\arguments{
\item{geo}{See \code{\link[soiltexture]{TT.plot}}}

\item{x}{A list with three items, "x", "y" and "z"}

\item{add}{If TRUE the image is added to an existing plot.}

\item{tri.sum.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.pos.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.tol}{See \code{\link[soiltexture]{TT.plot}}}

\item{unit.ps}{See \code{\link[soiltexture]{TT.plot}}}

\item{unit.tx}{See \code{\link[soiltexture]{TT.plot}}}

\item{b.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{l.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{main}{See \code{\link[soiltexture]{TT.plot}}}

\item{new.mar}{See \code{\link[soiltexture]{TT.plot}}}

\item{bg}{See \code{\link[soiltexture]{TT.plot}}}

\item{fg}{See \code{\link[soiltexture]{TT.plot}}}

\item{cex.main}{See \code{\link[soiltexture]{TT.plot}}}

\item{lang}{See \code{\link[soiltexture]{TT.plot}}}

\item{xlim}{Passed to \code{\link[graphics]{image}}}

\item{ylim}{Passed to \code{\link[graphics]{image}}}

\item{zlim}{Passed to \code{\link[graphics]{image}}}

\item{col}{Passed to \code{\link[graphics]{image}}}

\item{oldstyle}{Passed to \code{\link[graphics]{image}}}

\item{blr.clock}{See \code{\link[soiltexture]{TT.plot}}}

\item{tlr.an}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.tx}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.sum}{See \code{\link[soiltexture]{TT.plot}}}

\item{base.css.ps.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{\dots}{Additional parameters passed to image().}
}
\description{
A wrapper for the image() function adapted to texture triangles (plot
preparation). designed to plot the results of TT.mahalanobis() or TT.kde2d()
[to be written], before or after plot.
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
