\name{TT.image}
\alias{TT.image}
\title{Wrapper for the contour() function adapted to texture triangles.}
\description{A wrapper for the contour() function 
adapted to texture triangles (plot preparation).
designed to plot the results of TT.mahalanobis() or 
TT.kde2d() [to be written], before or after plot.}
\usage{TT.image(geo, x, add = FALSE, tri.sum.tst = NULL, tri.pos.tst = NULL, 
    text.tol = NULL, unit.ps = NULL, unit.tx = NULL, b.lim = NULL, 
    l.lim = NULL, main = NULL, new.mar = NULL, bg = NULL, fg = NULL, 
    cex.main = NULL, lang = NULL, xlim = NA, ylim = NA, zlim = NA, 
    col = rev(heat.colors(12)), oldstyle = FALSE, blr.clock = NULL, 
    tlr.an = NULL, blr.tx = NULL, text.sum = NULL, base.css.ps.lim = NULL, 
    ...)}
\arguments{
  \item{geo}{
}
  \item{x}{
}
  \item{add}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{text.tol}{
}
  \item{unit.ps}{
}
  \item{unit.tx}{
}
  \item{b.lim}{
}
  \item{l.lim}{
}
  \item{main}{
}
  \item{new.mar}{
}
  \item{bg}{
}
  \item{fg}{
}
  \item{cex.main}{
}
  \item{lang}{
}
  \item{xlim}{
}
  \item{ylim}{
}
  \item{zlim}{
}
  \item{col}{
}
  \item{oldstyle}{
}
  \item{blr.clock}{
}
  \item{tlr.an}{
}
  \item{blr.tx}{
}
  \item{text.sum}{
}
  \item{base.css.ps.lim}{
}
  \item{\dots}{Additional parameters passed to image().}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.}





