\name{TT.iwd}
\alias{TT.iwd}
\title{TT iwd}
\description{
}
\usage{
TT.iwd(tri.data, z.name, geo, css.names = NULL, tri.pol.data = NULL, text.tol = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, set.par = FALSE, n = 25, lims = c("points", "triangle")[1], max.dist = NULL, q.max.dist = 0.5, pow = 0.5)
}
\arguments{
  \item{tri.data}{
}
  \item{z.name}{
}
  \item{geo}{
}
  \item{css.names}{
}
  \item{tri.pol.data}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{set.par}{
}
  \item{n}{
}
  \item{lims}{
}
  \item{max.dist}{
}
  \item{q.max.dist}{
}
  \item{pow}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
