% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.kde2d}
\alias{TT.kde2d}
\title{Calculated the 2D probabilty density on an x-y grid.}
\usage{
TT.kde2d(
  geo,
  tri.data,
  css.names = NULL,
  text.tol = NULL,
  text.sum = NULL,
  blr.clock = NULL,
  tri.sum.tst = NULL,
  tri.pos.tst = NULL,
  set.par = FALSE,
  n = 25,
  lims = c("points", "triangle")[2]
)
}
\arguments{
\item{geo}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.data}{See \code{\link[soiltexture]{TT.plot}}}

\item{css.names}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.tol}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.sum}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.clock}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.sum.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.pos.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{set.par}{If TRUE parameters are set automatically to their defualt value}

\item{n}{Passed to \code{\link[MASS]{kde2d}}}

\item{lims}{Passed to \code{\link[MASS]{kde2d}}}
}
\description{
Function that calculated the 2D probabilty density on an x-y grid (and NOT
on the clay silt sand reference system). Wrapper around the kde2d function
from the MASS package.
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
