\name{TT.kde2d}
\alias{TT.kde2d}
\title{Calculated the 2D probabilty density on an x-y grid.}
\description{Function that calculated the 2D probabilty 
density on an x-y grid (and NOT on the clay silt sand 
reference system). Wrapper around the kde2d function from the 
MASS package.}
\usage{TT.kde2d(geo, tri.data, css.names = NULL, text.tol = NULL, text.sum = NULL, 
    blr.clock = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, 
    set.par = FALSE, n = 25, lims = c("points", "triangle")[2])}
\arguments{
  \item{geo}{
}
  \item{tri.data}{
}
  \item{css.names}{
}
  \item{text.tol}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{set.par}{
}
  \item{n}{
}
  \item{lims}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.}





