\name{TT.lines}
\alias{TT.lines}
\title{Internal. Used to plot line elements of a texture plot axis, ticks, arrows, etc.}
\description{Used to plot line elements of a texture plot axis, ticks, arrows, etc.}
\usage{TT.lines(geo = geo, at.1.s = TT.get("at"), at.2.s = 1 - TT.get("at"), 
    at.3.s = 0, at.1.e = TT.get("at"), at.2.e = 0, at.3.e = 1 - 
        TT.get("at"), text.tol = NULL, text.sum = NULL, blr.clock = NULL, 
    tri.sum.tst = NULL, tri.pos.tst = NULL)}
\arguments{
  \item{geo}{
}
  \item{at.1.s}{
}
  \item{at.2.s}{
}
  \item{at.3.s}{
}
  \item{at.1.e}{
}
  \item{at.2.e}{
}
  \item{at.3.e}{
}
  \item{text.tol}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





