\name{TT.locator}
\alias{TT.locator}
\title{Interactive (mouse clic) retrieval the CLAY SILT SAND coordinate of points on a texture triangle.}
\description{Interactive (mouse clic) retrieval the CLAY SILT SAND coordinate of points on a texture triangle. }
\usage{TT.locator(geo, css.names = NULL, text.tol = NULL, tri.sum.tst = NULL, 
    tri.pos.tst = FALSE, set.par = FALSE, n = 512, type = "n", 
    ...)}
\arguments{
  \item{geo}{
}
  \item{css.names}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{set.par}{
}
  \item{n}{
}
  \item{type}{
}
  \item{\dots}{Further argumets passed to locator() }
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.}





