\name{TT.mahalanobis}
\alias{TT.mahalanobis}
\title{TT mahalanobis}

\usage{
TT.mahalanobis(geo, tri.data, css.names = NULL, text.tol = NULL, text.sum = NULL, blr.clock = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, set.par = FALSE, n = 25, center = NULL, cov.mat = NULL, inverted = FALSE, ..., alr = FALSE, divisorvar = 2)
}
\arguments{
  \item{geo}{
}
  \item{tri.data}{
}
  \item{css.names}{
}
  \item{text.tol}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{set.par}{
}
  \item{n}{
}
  \item{center}{
}
  \item{cov.mat}{
}
  \item{inverted}{
}
  \item{\dots}{
}
  \item{alr}{
}
  \item{divisorvar}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





