\name{TT.normalise.sum}
\alias{TT.normalise.sum}
\title{Normalises the sum of the 3 particle size classes.}
\description{Normalises the sum of the 3 particle size classes in tri.data 
to text.sum (100\%).}
\usage{TT.normalise.sum(tri.data, css.names = NULL, text.sum = NULL, 
    text.tol = NULL, tri.pos.tst = NULL, residuals = FALSE)}
\arguments{
  \item{tri.data}{
}
  \item{css.names}{
}
  \item{text.sum}{
}
  \item{text.tol}{
}
  \item{tri.pos.tst}{
}
  \item{residuals}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.}





