\name{TT.normalise.sum.X}
\alias{TT.normalise.sum.X}
\title{Normalises the sum of the X particle size classes.}
\description{Normalises the sum of the X particle size classes
in tri.data to text.sum (100\%).}
\usage{
TT.normalise.sum.X(tri.data, text.sum = NULL, text.tol = NULL, tri.pos.tst = NULL, residuals = FALSE)
}
\arguments{
  \item{tri.data}{
}
  \item{text.sum}{
}
  \item{text.tol}{
}
  \item{tri.pos.tst}{
}
  \item{residuals}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





