\name{TT.par.op.set}
\alias{TT.par.op.set}
\title{Internal. Retrieve and set default values from options with default in "par()". }
\description{Retrieve and set default values from options with default in "par()"}
\usage{
TT.par.op.set(param, assign.op = TRUE, p.env = parent.frame())
}
\arguments{
  \item{param}{
}
  \item{assign.op}{
}
  \item{p.env}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





