\name{TT.phi2dia}
\alias{TT.phi2dia}
\title{Internal. Convert a soil particle phi value into diameter dia [micro-meters]. }
\description{Convert a soil particle phi value into diameter dia [micro-meters]. 
See also TT.dia2phi(). dia = (2^-phi)*1000. Not used by the package. }
\usage{TT.phi2dia(phi)}
\arguments{
  \item{phi}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





