% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.points}
\alias{TT.points}
\title{Plot a soil texture data table as points on an existing texture plot.}
\usage{
TT.points(
  tri.data,
  geo,
  css.names = NULL,
  z.name = NULL,
  base.css.ps.lim = NULL,
  dat.css.ps.lim = NULL,
  css.transf = NULL,
  text.transf.fun = NULL,
  trsf.add.opt1 = NULL,
  trsf.add.opt2 = NULL,
  text.tol = NULL,
  pch = NULL,
  fg = NULL,
  col = NULL,
  bg = NULL,
  cex = NULL,
  lwd = NULL,
  points.type = NULL,
  tri.sum.tst = NULL,
  tri.pos.tst = NULL,
  z.type = NULL,
  z.col.hue = NULL,
  z.cex.range = NULL,
  z.pch = NULL,
  text.sum = NULL,
  blr.clock = NULL,
  blr.tx = NULL
)
}
\arguments{
\item{tri.data}{See \code{\link[soiltexture]{TT.plot}}}

\item{geo}{See \code{\link[soiltexture]{TT.plot}}}

\item{css.names}{See \code{\link[soiltexture]{TT.plot}}}

\item{z.name}{See \code{\link[soiltexture]{TT.plot}}}

\item{base.css.ps.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{dat.css.ps.lim}{See \code{\link[soiltexture]{TT.plot}}}

\item{css.transf}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.transf.fun}{See \code{\link[soiltexture]{TT.plot}}}

\item{trsf.add.opt1}{See \code{\link[soiltexture]{TT.plot}}}

\item{trsf.add.opt2}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.tol}{See \code{\link[soiltexture]{TT.plot}}}

\item{pch}{See \code{\link[soiltexture]{TT.plot}}}

\item{fg}{See \code{\link[soiltexture]{TT.plot}}}

\item{col}{See \code{\link[soiltexture]{TT.plot}}}

\item{bg}{See \code{\link[soiltexture]{TT.plot}}}

\item{cex}{See \code{\link[soiltexture]{TT.plot}}}

\item{lwd}{See \code{\link[soiltexture]{TT.plot}}}

\item{points.type}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.sum.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{tri.pos.tst}{See \code{\link[soiltexture]{TT.plot}}}

\item{z.type}{See \code{\link[soiltexture]{TT.plot}}}

\item{z.col.hue}{See \code{\link[soiltexture]{TT.plot}}}

\item{z.cex.range}{See \code{\link[soiltexture]{TT.plot}}}

\item{z.pch}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.sum}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.clock}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.tx}{See \code{\link[soiltexture]{TT.plot}}}
}
\description{
Plot a soil texture data table as points on an existing texture plot.
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
