\name{TT.points}
\alias{TT.points}
\title{Plot a soil texture data table as points on an existing texture plot. }
\description{Plot a soil texture data table as points on an existing texture plot. }
\usage{TT.points(tri.data, geo, css.names = NULL, z.name = NULL, 
    base.css.ps.lim = NULL, dat.css.ps.lim = NULL, 
    css.transf = NULL, text.transf.fun = NULL, trsf.add.opt1 = NULL, 
    trsf.add.opt2 = NULL, text.tol = NULL, pch = NULL, 
    fg = NULL, col = NULL, bg = NULL, cex = NULL, lwd = NULL, 
    points.type = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, 
    z.type = NULL, z.col.hue = NULL, z.cex.range = NULL, 
    z.pch = NULL, text.sum = NULL, blr.clock = NULL, 
    blr.tx = NULL)}
\arguments{
  \item{tri.data}{
}
  \item{geo}{
}
  \item{css.names}{
}
  \item{z.name}{
}
  \item{base.css.ps.lim}{
}
  \item{dat.css.ps.lim}{
}
  \item{css.transf}{
}
  \item{text.transf.fun}{
}
  \item{trsf.add.opt1}{
}
  \item{trsf.add.opt2}{
}
  \item{text.tol}{
}
  \item{pch}{
}
  \item{fg}{
}
  \item{col}{
}
  \item{bg}{
}
  \item{cex}{
}
  \item{lwd}{
}
  \item{points.type}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{z.type}{
}
  \item{z.col.hue}{
}
  \item{z.cex.range}{
}
  \item{z.pch}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
  \item{blr.tx}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





