\name{TT.set}
\alias{TT.set}
\title{TT set}

\usage{
TT.set(..., reset = FALSE, par.list = "TT.par", bkp.par.list = "TT.par.bkp", par.env = TT.env)
}
\arguments{
  \item{\dots}{
}
  \item{reset}{
}
  \item{par.list}{
}
  \item{bkp.par.list}{
}
  \item{par.env}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





