% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.str}
\alias{TT.str}
\title{Stretch or reshape the range of value of some data set.}
\usage{
TT.str(x, str.min = 0, str.max = 1)
}
\arguments{
\item{x}{Vector of numeric values.}

\item{str.min}{Minimun value after stretching.}

\item{str.max}{Maximun value after stretching.}
}
\description{
Function to 'stretch' or reshape the range of value of some data set.
Usefull for cex parameter in plot.
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
