\name{TT.str}
\alias{TT.str}
\title{Internal. Stretch or reshape the range of value of some data set. }
\description{Function to 'stretch' or reshape the range of value of some data set. Usefull for cex parameter in plot. }
\usage{TT.str(x, str.min = 0, str.max = 1)}
\arguments{
  \item{x}{
}
  \item{str.min}{
}
  \item{str.max}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





