\name{TT.switch}
\alias{TT.switch}
\title{Internal. Used in the plot axis drawings.}
\description{Used in the plot axis drawings.}
\usage{TT.switch(blr.clock = TT.get("blr.clock"), c1 = NA, 
    c2 = NA, c3 = NA, c4 = NA, blr.order = c(1, 3, 
        2))}
\arguments{
  \item{blr.clock}{
}
  \item{c1}{
}
  \item{c2}{
}
  \item{c3}{
}
  \item{c4}{
}
  \item{blr.order}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





