\name{TT.text}
\alias{TT.text}
\title{Plot text labels for each values of a soil texture data table on an existing texture plot. }
\description{Plot text labels for each values of a soil texture data table on an existing texture plot. }
\usage{TT.text(tri.data, geo, labels = NULL, css.names = NULL, 
    base.css.ps.lim = NULL, dat.css.ps.lim = NULL, 
    css.transf = NULL, text.transf.fun = NULL, trsf.add.opt1 = NULL, 
    trsf.add.opt2 = NULL, text.tol = NULL, text.sum = NULL, 
    blr.clock = NULL, fg = NULL, col = NULL, cex = NULL, 
    font = NULL, family.op = NULL, adj = NULL, pos = NULL, 
    offset = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, 
    blr.tx = NULL)}
\arguments{
  \item{tri.data}{
}
  \item{geo}{
}
  \item{labels}{
}
  \item{css.names}{
}
  \item{base.css.ps.lim}{
}
  \item{dat.css.ps.lim}{
}
  \item{css.transf}{
}
  \item{text.transf.fun}{
}
  \item{trsf.add.opt1}{
}
  \item{trsf.add.opt2}{
}
  \item{text.tol}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
  \item{fg}{
}
  \item{col}{
}
  \item{cex}{
}
  \item{font}{
}
  \item{family.op}{
}
  \item{adj}{
}
  \item{pos}{
}
  \item{offset}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{blr.tx}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





