\name{TT.text.transf.X}
\alias{TT.text.transf.X}
\title{Log-linear transformation of a soil texture data table between 2 particle size systems (X classes).}
\description{Log-linear transformation of a soil texture data table 
('tri.data') from one 
particle size system ('dat.css.ps.lim') into another 
('base.css.ps.lim'). No limit in the number of partile size classes 
in the inputed and outputed texture tables. See TT.text.transf 
for transformation involving only 3 particle classes. 'tri.data' 
can only contain texture data.}
\usage{
TT.text.transf.X(tri.data,
base.ps.lim,
dat.ps.lim,
text.sum = NULL,
text.tol = NULL,
tri.sum.tst = NULL,
tri.pos.tst = NULL)
}
\arguments{
  \item{tri.data}{
}
  \item{base.ps.lim}{
}
  \item{dat.ps.lim}{
}
  \item{text.sum}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.}





