\name{TT.ticks.lab}
\alias{TT.ticks.lab}
\title{Internal. Plot the axis ticks' labels of a texture triangle plot. }
\description{Plot the axis ticks' labels of a texture triangle plot. }
\usage{TT.ticks.lab(geo, at = NULL, text.tol = NULL, text.sum = NULL, 
    blr.clock = NULL, tlr.an = NULL, tk.ls = NULL, 
    tri.sum.tst = NULL, tri.pos.tst = FALSE, col.axis = NULL, 
    font.axis = NULL, cex.axis = NULL, family.op = NULL)}
\arguments{
  \item{geo}{
}
  \item{at}{
}
  \item{text.tol}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
  \item{tlr.an}{
}
  \item{tk.ls}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{col.axis}{
}
  \item{font.axis}{
}
  \item{cex.axis}{
}
  \item{family.op}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





