\name{TT.ticks.lab}
\alias{TT.ticks.lab}
\title{TT ticks lab}
\description{
}
\usage{
TT.ticks.lab(geo, at = NULL, text.tol = NULL, tk.ls = NULL, tri.sum.tst = NULL, tri.pos.tst = FALSE, col.axis = NULL, font.axis = NULL, cex.axis = NULL, family.op = NULL)
}
\arguments{
  \item{geo}{
}
  \item{at}{
}
  \item{text.tol}{
}
  \item{tk.ls}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{col.axis}{
}
  \item{font.axis}{
}
  \item{cex.axis}{
}
  \item{family.op}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
