\name{TT.ticks}
\alias{TT.ticks}
\title{TT ticks}
\description{
}
\usage{
TT.ticks(geo, at = NULL, text.tol = NULL, tk.s = NULL, tri.sum.tst = NULL, tri.pos.tst = FALSE, lwd.axis = NULL, col.axis = NULL)
}
\arguments{
  \item{geo}{
}
  \item{at}{
}
  \item{text.tol}{
}
  \item{tk.s}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{lwd.axis}{
}
  \item{col.axis}{
}
}
\details{
}
\value{
}
\references{
}
\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}
\note{
}


\seealso{
}
\examples{}
