\name{TT.ticks}
\alias{TT.ticks}
\title{Internal. Plot the axis' ticks of a texture triangle plot. }
\description{Plot the axis' ticks of a texture triangle plot. }
\usage{TT.ticks(geo, at = NULL, text.tol = NULL, text.sum = NULL, blr.clock = NULL, 
    tk.s = NULL, tri.sum.tst = NULL, tri.pos.tst = FALSE, lwd.axis = NULL, 
    col.axis = NULL)}
\arguments{
  \item{geo}{
}
  \item{at}{
}
  \item{text.tol}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
  \item{tk.s}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{lwd.axis}{
}
  \item{col.axis}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





