\name{TT.vertices.plot}
\alias{TT.vertices.plot}
\title{Internal. Plot the vertices of a texture classification system. }
\description{Plot the vertices of a texture classification system, on top 
of an already drawn texture triangle plot. Also plot the 
vertices numbers. See TT.vertices.tbl() and TT.classes.tbl() 
for a non graphic, tabular equivalent of the plot.}
\usage{TT.vertices.plot(geo, class.sys = "HYPRES.TT", fg = NULL, 
    col = NULL, cex = NULL, font = NULL, family.op = NULL, 
    adj = NULL, pos = NULL, offset = NULL, blr.tx = NULL, 
    text.sum = NULL, blr.clock = NULL)}
\arguments{
  \item{geo}{
}
  \item{class.sys}{
}
  \item{fg}{
}
  \item{col}{
}
  \item{cex}{
}
  \item{font}{
}
  \item{family.op}{
}
  \item{adj}{
}
  \item{pos}{
}
  \item{offset}{
}
  \item{blr.tx}{
}
  \item{text.sum}{
}
  \item{blr.clock}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





