% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltexture.R
\name{TT.vertices.plot}
\alias{TT.vertices.plot}
\title{Plot the vertices of a texture classification system.}
\usage{
TT.vertices.plot(
  geo,
  class.sys = "HYPRES.TT",
  fg = NULL,
  col = NULL,
  cex = NULL,
  font = NULL,
  family.op = NULL,
  adj = NULL,
  pos = NULL,
  offset = NULL,
  blr.tx = NULL,
  text.sum = NULL,
  blr.clock = NULL
)
}
\arguments{
\item{geo}{See \code{\link[soiltexture]{TT.plot}}}

\item{class.sys}{See \code{\link[soiltexture]{TT.plot}}}

\item{fg}{See \code{\link[soiltexture]{TT.plot}}}

\item{col}{See \code{\link[soiltexture]{TT.plot}}}

\item{cex}{See \code{\link[soiltexture]{TT.plot}}}

\item{font}{See \code{\link[soiltexture]{TT.plot}}}

\item{family.op}{See \code{\link[soiltexture]{TT.plot}}}

\item{adj}{Passed to See \code{\link[soiltexture]{TT.text}}}

\item{pos}{Passed to See \code{\link[soiltexture]{TT.text}}}

\item{offset}{Passed to See \code{\link[soiltexture]{TT.text}}}

\item{blr.tx}{See \code{\link[soiltexture]{TT.plot}}}

\item{text.sum}{See \code{\link[soiltexture]{TT.plot}}}

\item{blr.clock}{See \code{\link[soiltexture]{TT.plot}}}
}
\description{
Plot the vertices of a texture classification system, on top of an already
drawn texture triangle plot. Also plot the vertices numbers. See
TT.vertices.tbl() and TT.classes.tbl() for a non graphic, tabular equivalent
of the plot.
}
\author{
Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb],
Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor
Zelazny [ctb], Rodolfo Marcondes Silva Souza [ctb], Jose Lucas Safanelli
[ctb], Alexandre ten Caten [ctb]
}
