\name{TT.vertices.tbl}
\alias{TT.vertices.tbl}
\title{Returns the table of vertices of a texture classification system.}
\description{Returns the table of vertices of a texture classification system. 
Returns the clay silt sand coordinates of each vertices. Use 
TT.classes.tbl() to know the vertices that bounds each texture 
class. See also TT.vertices.plot().}
\usage{TT.vertices.tbl(class.sys = "HYPRES.TT")}
\arguments{
  \item{class.sys}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





