\name{TT.xy.grid}
\alias{TT.xy.grid}
\title{Internal. Create a grid in the x-y coordinate system. }
\description{Create a grid in the x-y coordinate system. Most of the function 
is a reshaped extract from kde2d() from the MASS package, by 
Venables & Ripley (+ modifications)}
\usage{TT.xy.grid(x, y, n = 25)}
\arguments{
  \item{x}{
}
  \item{y}{
}
  \item{n}{
}
}



\author{Julien Moeys [aut, cre], Wei Shangguan [ctb], Rainer Petzold [ctb], Budiman Minasny [ctb], Bogdan Rosca [ctb], Nic Jelinski [ctb], Wiktor Zelazny [ctb],  Rodolfo Marcondes Silva Souza [ctb],  Jose Lucas Safanelli [ctb], Alexandre ten Caten [ctb]}





