\name{TT.xy2css}
\alias{TT.xy2css}
\title{TT xy2css}

\usage{
TT.xy2css(xy.data, geo, css.names = NULL, text.tol = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, set.par = FALSE)
}
\arguments{
  \item{xy.data}{
}
  \item{geo}{
}
  \item{css.names}{
}
  \item{text.tol}{
}
  \item{tri.sum.tst}{
}
  \item{tri.pos.tst}{
}
  \item{set.par}{
}
}



\author{Julien MOEYS <jules_m78-soiltexture@yahoo.fr>}





